+++ src/Admin/index.js
@@ -9,6 +9,24 @@ export const Admin = () => {
   const userData = localStorage.getItem("userData");
   const token = userData ? JSON.parse(userData).token : "invalid_token";
   const [authExpired, setAuthExpired] = useState(null);
+  const [categories, setCategories] = useState([]);
+  const [error, setError] = useState(false);
+  const categoriesRequest = async () => {
+    try {
+      const res = await fetch(`${BASE_URL}api/category`, {
+        method: "GET",
+        headers: {
+          "Content-Type": "application/json;charset=utf-8",
+        },
+      });
+
+      const data = await res.json();
+
+      setCategories(data);
+    } catch (error) {
+      setError(true);
+    }
+  };
   const checkAuth = async (token) => {
     try {
       const res = await fetch(`${BASE_URL}api/auth/`, {
@@ -21,17 +39,22 @@ export const Admin = () => {
 
       if (res.ok) {
         setAuthExpired(false);
+        categoriesRequest();
       } else {
         setAuthExpired(true);
       }
     } catch (error) {
-        setAuthExpired(true);
+      setAuthExpired(true);
     }
+  };
+
+  const updateSelector = (category) => {
+    setCategories([...categories, category]);
   };
 
   useEffect(() => {
     checkAuth(token);
-  }, [token]);
+  }, []);
 
   if (authExpired === null) return null;
   if (authExpired) return <Redirect to="/login" />;
@@ -39,9 +62,9 @@ export const Admin = () => {
   return (
     <>
       {console.log("render admin")}
-      <CategoryForm token={token}/>
-      <WordForm token={token}/>
-      <RegisterForm token={token}/>
+      <CategoryForm updateSelector={updateSelector} />
+      <WordForm categories={categories} error={error} />
+      <RegisterForm />
     </>
   );
 };
